iniciar a publicação de anuncio

+++ src/main/java/br/com/motur/dealbackendservice/core/dataproviders/repository/DealerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DealerRepository extends JpaRepository<DealerEntity, Long> {
+public interface DealerRepository extends JpaRepository<DealerEntity, Integer> {
 }
